fix: rename internal/p2p -> internal/ignore_p2p to ensure no package use it
