+++ internal/ignore_p2p/p2p_mocks.go
@@ -1,4 +1,4 @@
-package p2p
+package ignore_p2p
 
 //go:generate moq -pkg mocks -out ./mocks/peer_mock.go . PeerI
 //go:generate moq -pkg mocks -out ./mocks/message_handler_mock.go . MessageHandlerI
