+++ internal/ignore_p2p/tests/network_messenger_test.go
@@ -4,8 +4,8 @@ import (
 	"log/slog"
 	"testing"
 
-	"github.com/bitcoin-sv/arc/internal/p2p"
-	"github.com/bitcoin-sv/arc/internal/p2p/mocks"
+	p2p "github.com/bitcoin-sv/arc/internal/ignore_p2p"
+	"github.com/bitcoin-sv/arc/internal/ignore_p2p/mocks"
 	"github.com/libsv/go-p2p/chaincfg/chainhash"
 	"github.com/libsv/go-p2p/wire"
 	"github.com/stretchr/testify/require"
